# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.DocumentType:
    enum:
    - rich_text
    - file
    type: string
    x-enum-varnames:
    - DocumentTypeRichText
    - DocumentTypeFile
...
  handler.createDocumentRequest:
    properties:
      docType:
        $ref: '#/definitions/domain.DocumentType'
      parentId:
        type: string
      sourceStorageKey:
        type: string
      summary:
        type: string
      title:
        type: string
    required:
    - title
    type: object
...
  handler.moveDocumentRequest:
    properties:
      parentId:
        type: string
      sortOrder:
        type: integer
    type: object
...
  handler.updateDocumentRequest:
    properties:
      sourceStorageKey:
        type: string
      summary:
        type: string
      title:
        type: string
    type: object
...
  response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: 科研项目文档管理系统后端 API。
  title: Research Admin Backend API
  version: "1.0"
paths:
...
  /documents/{documentId}:
    delete:
      description: 软删除文档。
      parameters:
      - description: 文档 ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 删除文档
      tags:
      - documents
    get:
      description: 返回单个文档元数据和当前用户最终权限。
      parameters:
      - description: 文档 ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 获取文档详情
      tags:
      - documents
    patch:
      consumes:
      - application/json
      description: 更新文档元数据，不处理正文内容。
      parameters:
      - description: 文档 ID
        in: path
        name: documentId
        required: true
        type: string
      - description: 文档更新参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updateDocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 更新文档元数据
      tags:
      - documents
...
  /documents/{documentId}/archive:
    post:
      description: 将文档标记为 archived。
      parameters:
      - description: 文档 ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 归档文档
      tags:
      - documents
  /documents/{documentId}/download:
    get:
      description: 返回文件型文档的存储引用。
      parameters:
      - description: 文档 ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 下载文件型文档
      tags:
      - documents
  /documents/{documentId}/move:
    post:
      consumes:
      - application/json
      description: 移动文档到新的父节点并更新排序。
      parameters:
      - description: 文档 ID
        in: path
        name: documentId
        required: true
        type: string
      - description: 移动参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.moveDocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 移动文档
      tags:
      - documents
...
  /documents/{documentId}/restore:
    post:
      description: 将归档文档恢复为 active。
      parameters:
      - description: 文档 ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 恢复文档
      tags:
      - documents
...
  /workspaces/{workspaceId}/documents:
    post:
      consumes:
      - application/json
      description: 在 workspace 下创建文档元数据。根文档要求 workspace 成员权限，子文档要求父文档 EDIT 权限。
      parameters:
      - description: Workspace ID
        in: path
        name: workspaceId
        required: true
        type: string
      - description: 文档参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createDocumentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 创建文档
      tags:
      - documents
  /workspaces/{workspaceId}/documents/upload:
    post:
      consumes:
      - multipart/form-data
      description: 创建文件型文档元数据，真实文件存储由后续存储模块接入。
      parameters:
      - description: Workspace ID
        in: path
        name: workspaceId
        required: true
        type: string
      - description: 上传文件
        in: formData
        name: file
        required: true
        type: file
      - description: 父文档 ID
        in: formData
        name: parentId
        type: string
      - description: 文档标题
        in: formData
        name: title
        type: string
      - description: 文档摘要
        in: formData
        name: summary
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 上传文件型文档
      tags:
      - documents
...
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"